backend_python/requirements.txt
fastapi
uvicorn
pymongo
motor
python-dotenv
pydantic
pydantic-settings
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
websockets==12.0
bcrypt==3.2.0
python-multipart
email-validator
